text-to-sql-agent/cloudera/cml_requirements.txt
# CML Model Requirements
# Optimized for Cloudera Machine Learning environment

# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM and AI
openai==1.12.0
anthropic==0.18.1

# Vector stores
chromadb==0.4.22
sentence-transformers==2.3.1

# Database
sqlalchemy==2.0.25
pyhive==0.7.0
sqlparse==0.4.4
thrift==0.16.0
sasl==0.3.1
thrift-sasl==0.4.3

# API (minimal for model serving)
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Visualization
pandas==2.2.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.2

# Utilities
tenacity==8.2.3
jinja2==3.1.3
loguru==0.7.2
rich==13.7.0

# CML specific
cmlapi==0.0.10
